requirements.txt
Django==1.11.8  
djangorestframework==3.7.3
markdown 
django-filter 
django-cors-headers==2.1.0